+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/prompt_guard/models/AuditDataActivityConfig.java
@@ -1,0 +1,29 @@
+package cloud.pangeacyber.pangea.prompt_guard.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Builder;
+import lombok.NonNull;
+import lombok.Value;
+import lombok.extern.jackson.Jacksonized;
+
+@Builder
+@Jacksonized
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(Include.NON_NULL)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+@Value
+public final class AuditDataActivityConfig {
+
+	@NonNull
+	Boolean enabled;
+
+	@NonNull
+	String auditServiceConfigId;
+
+	@NonNull
+	AuditDataActivityConfigAreas areas;
+}